Support latest faraday
Current master branch can't pass tests due to Faraday's changes.
https://travis-ci.org/cookpad/garage_client/builds/210856611

Faraday::Env inherites Faraday::Options and its #merge! is different
from Hash#merge!. The method does not use given block and failed
merging when a value in `@env` with a key (A) is not nil but another
value in `response.env` with same key (A) is nil with NomethodError
on nil. The cause of the degression might be brought by
https://github.com/lostisland/faraday/pull/662 but I'm not sure now.
So for now, just work-around by using more primitive method, #each,
instread.